+++ src/modules/claim-link/types/index.ts
@@ -3,38 +3,35 @@ import TRedeem from './redeem'
 import TDepositWithAuthorization from './deposit-with-authorization'
 import TConstructorArgs from './constructor-args' 
 import TUpdateAmount from './update-amount'
-import TGetNextTransferId from './get-next-transfer-id'
 import TGenerateClaimUrl from './generate-claim-url'
 import TDefineDomain from './define-domain'
 import TGetStatus from './get-status'
 import TDeposit from './deposit'
-import TInitialize from './initialize'
+import TDefineValue from './define-value'
 
 interface IClaimLink {
   _getCurrentFee: TGetCurrentFee
   redeem: TRedeem
   depositWithAuthorization: TDepositWithAuthorization
   updateAmount: TUpdateAmount
-  _getNextTransferId: TGetNextTransferId
   _defineDomain: TDefineDomain
   generateClaimUrl: TGenerateClaimUrl
   getStatus: TGetStatus
   deposit: TDeposit
-  initialize: TInitialize
+  _defineValue: TDefineValue
 }
 
 export {
   TGetCurrentFee,
   TRedeem,
-  TInitialize,
   TDepositWithAuthorization,
   TConstructorArgs,
   TUpdateAmount,
-  TGetNextTransferId,
   TGenerateClaimUrl,
   TDefineDomain,
   TGetStatus,
-  TDeposit
+  TDeposit,
+  TDefineValue
 }
 
 export default IClaimLink