+++ src/modules/claim-link/types/initialize/index.ts
@@ -1,3 +1,0 @@
-type TInitialize = () => Promise<void>
-
-export default TInitialize